refactor: Organized client project structure
