+++ client/src/components/editor/EditorLoader.tsx
