+++ client/src/components/shared/card/ErrorCard.tsx
