+++ client/src/layouts/RootLayout.tsx
@@ -1,4 +1,4 @@
-import HomeSideBar from "@/components/HomeSideBar";
+import HomeSideBar from "@/components/shared/nav/HomeSideBar";
 import { Toaster } from "@/shadcn/components/ui/toaster";
 import { Outlet } from "react-router";
