+++ client/src/screens/EditorScreen.tsx
@@ -1,11 +1,11 @@
-import Menu from "@/components/Menu";
+import Menu from "@/components/shared/nav/Menu";
 import Editor from "@/components/editor/Editor";
 import { useState } from "react";
 import { EditorContext } from "@/context/EditorContext";
 import { useParams } from "react-router";
 import { Test } from "@/models/Program";
 import { useTest } from "@/hooks/useTest";
-import { ErrorCard } from "@/components/ErrorCard";
+import { ErrorCard } from "@/components/shared/card/ErrorCard";
 
 function EditorScreen() {
   const { testId } = useParams();
