+++ client/src/screens/LoginScreen.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/LoginForm";
+import { LoginForm } from "@/components/auth/LoginForm";
 
 export default function LoginScreen() {
   return (
