+++ client/src/screens/SignUpScreen.tsx
@@ -1,4 +1,4 @@
-import { SignUpForm } from "@/components/SignUpForm";
+import { SignUpForm } from "@/components/auth/SignUpForm";
 
 export default function SignUpScreen() {
   return (
